pyproject.toml
[project]
name = "gen-ai-capstone-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=0.6.3",
    "openai>=1.60.1",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-environ>=0.4.54",
    "sentence-transformers>=3.4.0",
    "streamlit>=1.41.1",
]
